requirements.txt
Flask
argon2-cffi
pyotp
qrcode[pil]>=7.0
webauthn
pycryptodome
